[FEAT] Date 데이터 정제하는 함수들 모으기 #NAA-60
